+++ src/configs/index.ts
@@ -1,3 +1,11 @@
+export { base } from './base';
 export { minimal } from './minimal';
 export { recommended } from './recommended';
 export { strict } from './strict';
+
+export {
+  minimalRules,
+  recommendedRules,
+  strictRules,
+  type TailwindcssRules,
+} from './rules';